adjusted repo to new msg types

+++ rotors_gazebo_plugins/src/gazebo_motor_model.cpp
@@ -131,7 +131,7 @@ void GazeboMotorModel::VelocityCallback(const mav_msgs::MotorSpeedPtr& rot_veloc
   CHECK(rot_velocities->motor_speed.size() > motor_number_)
       << "You tried to access index " << motor_number_
       << " of the MotorSpeed message array which is of size " << rot_velocities->motor_speed.size() << ".";
-  ref_motor_rot_vel_ = std::min(rot_velocities->motor_speed[motor_number_], static_cast<float>(max_rot_velocity_));
+  ref_motor_rot_vel_ = std::min(rot_velocities->motor_speed[motor_number_], static_cast<double>(max_rot_velocity_));
 }
 
 void GazeboMotorModel::UpdateForcesAndMoments() {
